requirements.txt
anyio==4.11.0
azure-core==1.36.0
azure-data-tables==12.7.0
azure-functions==1.24.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cssselect==1.3.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
isodate==0.7.2
jieba3k==0.35.1
joblib==1.5.2
lxml==6.0.2
lxml_html_clean==0.4.3
MarkupSafe==3.0.3
multidict==6.7.0
newspaper3k==0.2.8
nltk==3.9.2
packaging==25.0
pillow==12.0.0
propcache==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.22.0
